+++ src/db-up/azext_db_up/commands.py
@@ -5,13 +5,13 @@
 
 from azure.cli.core.commands import CliCommandType
 from azure.cli.command_modules.rdbms._client_factory import cf_mysql_servers
-from azext_rdbms_up._validators import process_mysql_namespace
-from azext_rdbms_up._transformers import table_transform_connection_string
+from azext_db_up._validators import process_mysql_namespace
+from azext_db_up._transformers import table_transform_connection_string
 
 
 def load_command_table(self, _):  # pylint: disable=too-many-locals, too-many-statements
     mysql_servers_sdk = CliCommandType(
-        operations_tmpl='azext_rdbms_up.vendored_sdks.azure_mgmt_rdbms.mysql.operations.servers_operations'
+        operations_tmpl='azext_db_up.vendored_sdks.azure_mgmt_rdbms.mysql.operations.servers_operations'
                         '#ServersOperations.{}',
         client_factory=cf_mysql_servers
     )