+++ src/db-up/HISTORY.rst
@@ -6,6 +6,7 @@
 0.1.1 (tbd)
 +++++++++++++++++
 * `az mysql up`- create a resource group if a name is provided that is not an existing one
+* `az mysql up`- changes to output to only show connection details and enable table format
 
 0.1.0 (2019-1-23)
 +++++++++++++++++